reorder storage interface a bit to reduce code duplication
