+++ src/ZArray.jl
@@ -84,14 +84,11 @@ end
 # A lot of the JSON parsing should be moved to a function, since
 # this will be the same for other backends
 function ZArray(s::T, mode="r") where T <: AbstractStore
-    jsonstr = readmeta(s, ".zarray")
-    metadata = Metadata(jsonstr)
-    attrs = getattrs(s)
-    writeable = mode == "w"
-    ZArray{eltype(metadata), length(metadata.shape), typeof(metadata.compressor), T}(
-        metadata, s, attrs, writeable)
-    z = ZArray(
-        metadata, s, attrs, writeable)
+  metadata = getmetadata(s)
+  attrs = getattrs(s)
+  writeable = mode == "w"
+  ZArray{eltype(metadata), length(metadata.shape), typeof(metadata.compressor), T}(
+    metadata, s, attrs, writeable)
 end
 
 """
@@ -191,9 +188,8 @@ function readblock!(aout, z::ZArray{<:Any, N}, r::CartesianIndices{N}; readmode=
         # Extract them as CartesianIndices objects
         i_in_a = CartesianIndices(map(i -> i[1], ii))
         i_in_out = CartesianIndices(map(i -> i[2], ii))
-
         # Uncompress a chunk
-        if readmode || (isinitialized(z.storage, bI + one(bI)) && (size(i_in_a) != size(i_in_a)))
+        if readmode || (isinitialized(z.storage, bI + one(bI)) && (size(i_in_a) != size(a)))
           readchunk!(maybeinner(a), z, bI + one(bI))
         end
 
@@ -264,11 +260,11 @@ Read the chunk specified by `i` from the Zarray `z` and write its content to `a`
 """
 function readchunk!(a::DenseArray,z::ZArray{<:Any,N},i::CartesianIndex{N}) where N
     length(a) == prod(z.metadata.chunks) || throw(DimensionMismatch("Array size does not equal chunk size"))
-    curchunk = getchunk(z.storage, i)
+    curchunk = z.storage[i]
     if curchunk === nothing
         fill!(a, z.metadata.fill_value)
     else
-        read_uncompress!(a, curchunk, z.metadata.compressor, z.storage)
+        zuncompress(a, curchunk, z.metadata.compressor)
     end
     a
 end
@@ -281,11 +277,9 @@ Write the data from the array `a` to the chunk `i` in the ZArray `z`
 function writechunk!(a::DenseArray, z::ZArray{<:Any,N}, i::CartesianIndex{N}) where N
     z.writeable || error("ZArray not in write mode")
     length(a) == prod(z.metadata.chunks) || throw(DimensionMismatch("Array size does not equal chunk size"))
-    curchunk = getchunk(z.storage, i)
-    if curchunk === nothing
-      curchunk = createchunk(z.storage,i)
-    end
-    write_compress(a, curchunk, z.metadata.compressor)
+    dtemp = UInt8[]
+    zcompress(a,dtemp,z.metadata.compressor)
+    z.storage[i]=dtemp
     a
 end
